+++ metadata.rb
@@ -9,3 +9,5 @@ version '1.6.1'
 supports 'ubuntu'
 supports 'redhat'
 supports 'centos'
+
+depends 'chef-sugar'